Polars update ci (#1433)
* update ci to run tests on polars-dev PRs


* fix type


* use Union type


* update deps


* fix lint


* loosen pylint contraints, fix strategies circular import


---------


update req files
