+++ .github/workflows/ci-tests.yml
@@ -10,6 +10,7 @@ on:
     branches:
       - main
       - dev
+      - polars-dev
       - bugfix
       - "release/*"
