+++ ci/requirements-py3.10-pandas1.5.3-pydantic1.10.11.txt
@@ -129,7 +129,7 @@ grpcio==1.58.0
     # via ray
 h11==0.14.0
     # via uvicorn
-hypothesis==6.98.9
+hypothesis==6.98.10
 identify==2.5.29
     # via pre-commit
 idna==3.4
@@ -296,6 +296,7 @@ pandas==1.5.3
     #   modin
     #   partd
     #   petl
+    #   polars
     #   pyspark
     #   ray
 pandas-stubs==1.5.2.221213
@@ -320,6 +321,7 @@ platformdirs==3.10.0
     #   virtualenv
 pluggy==1.3.0
     # via pytest
+polars==0.20.10
 pre-commit==3.4.0
 prometheus-client==0.17.1
     # via jupyter-server
@@ -340,6 +342,7 @@ pydantic==1.10.11
     # via
     #   fastapi
     #   modin
+    #   polars
     #   ray
 pygments==2.16.1
     # via
