+++ ci/requirements-py3.10-pandas2.2.0-pydantic1.10.11.txt
@@ -292,6 +292,7 @@ pandas==2.2.0
     #   modin
     #   partd
     #   petl
+    #   polars
     #   pyspark
     #   ray
 pandas-stubs==2.2.0.240218
@@ -314,6 +315,7 @@ platformdirs==4.2.0
     #   virtualenv
 pluggy==1.4.0
     # via pytest
+polars==0.20.10
 pre-commit==3.6.2
 prometheus-client==0.20.0
     # via jupyter-server
@@ -333,6 +335,7 @@ pycparser==2.21
 pydantic==1.10.11
     # via
     #   fastapi
+    #   polars
     #   ray
 pygments==2.17.2
     # via
