+++ ci/requirements-py3.9-pandas1.5.3-pydantic2.3.0.txt
@@ -131,7 +131,7 @@ grpcio==1.58.0
     # via ray
 h11==0.14.0
     # via uvicorn
-hypothesis==6.98.9
+hypothesis==6.98.10
 identify==2.5.29
     # via pre-commit
 idna==3.4
@@ -305,6 +305,7 @@ pandas==1.5.3
     #   modin
     #   partd
     #   petl
+    #   polars
     #   pyspark
     #   ray
 pandas-stubs==1.5.2.221213
@@ -329,6 +330,7 @@ platformdirs==3.10.0
     #   virtualenv
 pluggy==1.3.0
     # via pytest
+polars==0.20.10
 pre-commit==3.4.0
 prometheus-client==0.17.1
     # via jupyter-server
@@ -349,6 +351,7 @@ pydantic==2.3.0
     # via
     #   fastapi
     #   modin
+    #   polars
     #   ray
 pydantic-core==2.6.3
     # via pydantic
