+++ ci/requirements-py3.9-pandas2.2.0-pydantic1.10.11.txt
@@ -299,6 +299,7 @@ pandas==2.2.0
     #   modin
     #   partd
     #   petl
+    #   polars
     #   pyspark
     #   ray
 pandas-stubs==2.2.0.240218
@@ -321,6 +322,7 @@ platformdirs==4.2.0
     #   virtualenv
 pluggy==1.4.0
     # via pytest
+polars==0.20.10
 pre-commit==3.6.2
 prometheus-client==0.20.0
     # via jupyter-server
@@ -340,6 +342,7 @@ pycparser==2.21
 pydantic==1.10.11
     # via
     #   fastapi
+    #   polars
     #   ray
 pygments==2.17.2
     # via
