+++ ci/requirements-py3.9-pandas2.2.0-pydantic2.3.0.txt
@@ -301,6 +301,7 @@ pandas==2.2.0
     #   modin
     #   partd
     #   petl
+    #   polars
     #   pyspark
     #   ray
 pandas-stubs==2.2.0.240218
@@ -323,6 +324,7 @@ platformdirs==4.2.0
     #   virtualenv
 pluggy==1.4.0
     # via pytest
+polars==0.20.10
 pre-commit==3.6.2
 prometheus-client==0.20.0
     # via jupyter-server
@@ -342,6 +344,7 @@ pycparser==2.21
 pydantic==2.3.0
     # via
     #   fastapi
+    #   polars
     #   ray
 pydantic-core==2.6.3
     # via pydantic
