+++ pandera/backends/polars/checks.py
@@ -38,7 +38,7 @@ class PolarsCheckBackend(BaseCheckBackend):
         """Implements aggregation behavior for check object."""
         raise NotImplementedError
 
-    def preprocess(self, check_obj: pl.LazyFrame, key: str):
+    def preprocess(self, check_obj: pl.LazyFrame, key: Optional[str]):
         """Preprocesses a check object before applying the check function."""
         # This handles the case of Series validation, which has no other context except
         # for the index to groupby on. Right now grouping by the index is not allowed.
