+++ requirements-docs.txt
@@ -1,5 +1,5 @@
 #
-# This file is autogenerated by pip-compile with Python 3.11
+# This file is autogenerated by pip-compile with Python 3.10
 # by the following command:
 #
 #    pip-compile --no-emit-index-url --output-file=requirements-docs.txt requirements.in
@@ -53,9 +53,7 @@ certifi==2023.7.22
     #   pyproj
     #   requests
 cffi==1.15.1
-    # via
-    #   argon2-cffi-bindings
-    #   cryptography
+    # via argon2-cffi-bindings
 cfgv==3.4.0
     # via pre-commit
 chardet==5.2.0
@@ -89,11 +87,7 @@ colorlog==6.7.0
 commonmark==0.9.1
     # via recommonmark
 coverage[toml]==7.3.1
-    # via
-    #   coverage
-    #   pytest-cov
-cryptography==42.0.2
-    # via secretstorage
+    # via pytest-cov
 dask==2023.9.2
     # via
     #   -r requirements.in
@@ -180,10 +174,6 @@ isort==5.12.0
     #   pylint
 jaraco-classes==3.3.0
     # via keyring
-jeepney==0.8.0
-    # via
-    #   keyring
-    #   secretstorage
 jinja2==3.1.2
     # via
     #   distributed
@@ -480,8 +470,6 @@ rpds-py==0.10.3
     #   referencing
 scipy==1.11.2
     # via -r requirements.in
-secretstorage==3.3.3
-    # via keyring
 send2trash==1.8.2
     # via jupyter-server
 shapely==2.0.1
@@ -558,6 +546,13 @@ text-unidecode==1.3
     # via python-slugify
 tinycss2==1.2.1
     # via nbconvert
+tomli==2.0.1
+    # via
+    #   black
+    #   coverage
+    #   mypy
+    #   pylint
+    #   pytest
 tomlkit==0.12.1
     # via pylint
 toolz==0.12.0
@@ -585,9 +580,7 @@ twine==4.0.2
 typeguard==4.1.5
     # via -r requirements.in
 typer[all]==0.9.0
-    # via
-    #   frictionless
-    #   typer
+    # via frictionless
 types-click==7.1.8
     # via -r requirements.in
 types-pkg-resources==0.1.3
@@ -605,6 +598,8 @@ types-urllib3==1.26.25.14
 typing-extensions==4.8.0
     # via
     #   -r requirements.in
+    #   astroid
+    #   black
     #   fastapi
     #   mypy
     #   pydantic
@@ -612,6 +607,7 @@ typing-extensions==4.8.0
     #   typeguard
     #   typer
     #   typing-inspect
+    #   uvicorn
 typing-inspect==0.9.0
     # via -r requirements.in
 tzdata==2023.3
